Backport https://github.com/reicast/reicast-emulator/commit/4bd23f2c8e2ac9e15a78cd6a21c31328c6492829

+++ core/rend/gl4/gles.cpp
@@ -374,7 +374,7 @@ void main() \n\
       Pixel pixel; \n\
       pixel.color = color; \n\
       pixel.depth = gl_FragDepth; \n\
-      pixel.seq_num = pp_Number; \n\
+      pixel.seq_num = uint(pp_Number); \n\
       pixel.next = imageAtomicExchange(abufferPointerImg, coords, idx); \n\
       pixels[idx] = pixel; \n\
       \n\
